+++ src/main/java/com/example/pidev1/Entity/Department.java
@@ -1,4 +1,4 @@
-package com.example.pidev1.entity;
+package com.example.pidev1.Entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,22 +7,23 @@ import lombok.Setter;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.util.Set;
 
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table( name = "Rdv")
-public class Rdv implements Serializable {
+@Table( name = "Department")
+public class Department implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="idRdv")
-    private Long idRdv;
+    @Column(name="idDepartment")
+    private Long idDepartment;
     private String Name;
 
-    @ManyToOne
-    Student studentt;
+    @ManyToMany(mappedBy="departments", cascade = CascadeType.ALL)
+    private Set<Student> studentSet;
 
 
 }
