+++ src/main/java/com/example/pidev1/Entity/Follow_upRequest.java
@@ -1,4 +1,4 @@
-package com.example.pidev1.entity;
+package com.example.pidev1.Entity;
 
 public enum Follow_upRequest {
     Accepte,Refusee,In_progress
