+++ src/main/java/com/example/pidev1/Entity/Participation.java
@@ -1,0 +1,29 @@
+package com.example.pidev1.Entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.util.Set;
+
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Table( name = "Participation")
+public class Participation implements Serializable {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="idParticipation")
+    private long idParticipation;
+    private String nameP;
+    private String EmailP;
+
+    @ManyToMany(cascade = CascadeType.ALL)
+    private Set<Event> events;
+}
