+++ src/main/java/com/example/pidev1/Entity/Photo.java
@@ -1,0 +1,24 @@
+package com.example.pidev1.Entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class Photo {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String url;
+
+    @OneToOne(mappedBy = "photo")
+    private Event event;
+}
