+++ src/main/java/com/example/pidev1/Entity/Rdv.java
@@ -1,4 +1,4 @@
-package com.example.pidev1.entity;
+package com.example.pidev1.Entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,23 +7,22 @@ import lombok.Setter;
 
 import javax.persistence.*;
 import java.io.Serializable;
-import java.util.Set;
 
 @Entity
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table( name = "Department")
-public class Department implements Serializable {
+@Table( name = "Rdv")
+public class Rdv implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="idDepartment")
-    private Long idDepartment;
+    @Column(name="idRdv")
+    private Long idRdv;
     private String Name;
 
-    @ManyToMany(mappedBy="departments", cascade = CascadeType.ALL)
-    private Set<Student> studentSet;
+    @ManyToOne
+    Student studentt;
 
 
 }
