+++ src/main/java/com/example/pidev1/Entity/Role.java
@@ -1,4 +1,4 @@
-package com.example.pidev1.entity;
+package com.example.pidev1.Entity;
 
 public enum Role {
     Teacher,Administration,RH
