+++ src/main/java/com/example/pidev1/Repository/EmployersRepo.java
@@ -1,6 +1,6 @@
-package com.example.pidev1.repository;
+package com.example.pidev1.Repository;
 
-import com.example.pidev1.entity.Employers;
+import com.example.pidev1.Entity.Employers;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
