+++ src/main/java/com/example/pidev1/Repository/ParticipationRepository.java
@@ -1,0 +1,4 @@
+package com.example.pidev1.Repository;
+
+public interface ParticipationRepository {
+}
