+++ src/main/java/com/example/pidev1/Repository/PhotoRepository.java
@@ -1,0 +1,9 @@
+package com.example.pidev1.Repository;
+
+import com.example.pidev1.Entity.Photo;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PhotoRepository extends JpaRepository<Photo,Long> {
+}
