+++ src/main/java/com/example/pidev1/Repository/StudentRepository.java
@@ -1,0 +1,7 @@
+package com.example.pidev1.Repository;
+
+import com.example.pidev1.Entity.Student;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudentRepository extends JpaRepository<Student,Long> {
+}
