+++ src/main/java/com/example/pidev1/Service/IHost.java
@@ -1,0 +1,4 @@
+package com.example.pidev1.Service;
+
+public interface IHost {
+}
