+++ src/main/java/com/example/pidev1/Service/IParticipation.java
@@ -1,0 +1,4 @@
+package com.example.pidev1.Service;
+
+public interface IParticipation {
+}
