+++ src/main/java/com/example/pidev1/Service/ParticipationService.java
@@ -1,0 +1,9 @@
+package com.example.pidev1.Service;
+
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class ParticipationService implements IParticipation {
+}
