--- src/main/java/com/example/pidev1/entity/Event.java
@@ -1,36 +1,0 @@
-package com.example.pidev1.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.io.Serializable;
-import java.util.Date;
-import java.util.Set;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Table( name = "Event")
-public class Event implements Serializable {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="idEvent")
-    private Long idEvent;
-    private Long nbr_of_places;
-    private String place;
-    private String Organisation;
-    @Temporal(TemporalType.DATE)
-    private Date date;
-    @Enumerated(EnumType.STRING)
-    private EventType Type;
-
-    @ManyToMany(mappedBy="events", cascade = CascadeType.ALL)
-    private Set<Student> students;
-
-
-}
