--- src/main/java/com/example/pidev1/entity/EventType.java
@@ -1,5 +1,0 @@
-package com.example.pidev1.entity;
-
-public enum EventType {
-    App0,Music,Sport
-}
